+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ## [Unreleased]
 ### Added
+- install script of linting pre-commit hook for kotlin and dotnet on Mac